# supervisor-console/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zippy Logitech API",
    "description": "Logistics Business Intelligence Platform API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/vehicles": {
      "post": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Create Vehicle",
        "operationId": "create_vehicle_api_v1_vehicles_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "List Vehicles",
        "description": "List all vehicles with optional filters",
        "operationId": "list_vehicles_api_v1_vehicles_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category (LCV, ICV, HCV, Tipper)",
              "title": "Category"
            },
            "description": "Filter by category (LCV, ICV, HCV, Tipper)"
          },
          {
            "name": "manufacturer",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by manufacturer",
              "title": "Manufacturer"
            },
            "description": "Filter by manufacturer"
          },
          {
            "name": "min_payload",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum payload in kg",
              "title": "Min Payload"
            },
            "description": "Minimum payload in kg"
          },
          {
            "name": "max_payload",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum payload in kg",
              "title": "Max Payload"
            },
            "description": "Maximum payload in kg"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "description": "Maximum number of results",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Offset for pagination",
              "default": 0,
              "title": "Offset"
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/{vehicle_id}": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Get Vehicle",
        "description": "Get a specific vehicle by ID",
        "operationId": "get_vehicle_api_v1_vehicles__vehicle_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vehicle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Update Vehicle",
        "operationId": "update_vehicle_api_v1_vehicles__vehicle_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Vehicle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/category/{category}": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Get Vehicles By Category",
        "description": "Get vehicles by category",
        "operationId": "get_vehicles_by_category_api_v1_vehicles_category__category__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/manufacturer/{manufacturer}": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Get Vehicles By Manufacturer",
        "description": "Get vehicles by manufacturer",
        "operationId": "get_vehicles_by_manufacturer_api_v1_vehicles_manufacturer__manufacturer__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "manufacturer",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Manufacturer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehicleListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vehicles/recommend": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Recommend Vehicles",
        "description": "Recommend suitable vehicles based on cargo requirements.\nReturns vehicles with payload capacity >= required weight (with 20% buffer).",
        "operationId": "recommend_vehicles_api_v1_vehicles_recommend_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "weight_kg",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0.0,
              "description": "Cargo weight in kg",
              "title": "Weight Kg"
            },
            "description": "Cargo weight in kg"
          },
          {
            "name": "volume_cbm",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "exclusiveMinimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Cargo volume in cubic meters",
              "title": "Volume Cbm"
            },
            "description": "Cargo volume in cubic meters"
          },
          {
            "name": "distance_km",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "exclusiveMinimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "Transport distance in km",
              "title": "Distance Km"
            },
            "description": "Transport distance in km"
          },
          {
            "name": "category_preference",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Preferred category",
              "title": "Category Preference"
            },
            "description": "Preferred category"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10,
              "description": "Number of recommendations",
              "default": 5,
              "title": "Limit"
            },
            "description": "Number of recommendations"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehicleRecommendResponse"
                  },
                  "title": "Response Recommend Vehicles Api V1 Vehicles Recommend Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/manufacturers": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "List Manufacturers",
        "description": "Get list of all manufacturers",
        "operationId": "list_manufacturers_api_v1_manufacturers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/categories": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "List Categories",
        "description": "Get list of all vehicle categories",
        "operationId": "list_categories_api_v1_categories_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VehicleCreate": {
        "properties": {
          "manufacturer": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Manufacturer"
          },
          "model_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Model Name"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "body_type": {
            "type": "string",
            "title": "Body Type",
            "default": "open"
          },
          "gvw_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gvw Kg"
          },
          "payload_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Kg"
          },
          "mileage_kmpl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mileage Kmpl"
          },
          "price_ex_showroom": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Ex Showroom"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "manufacturer",
          "model_name",
          "category"
        ],
        "title": "VehicleCreate"
      },
      "VehicleListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "vehicles": {
            "items": {
              "$ref": "#/components/schemas/VehicleResponse"
            },
            "type": "array",
            "title": "Vehicles"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "vehicles"
        ],
        "title": "VehicleListResponse",
        "description": "Paginated list of vehicles"
      },
...
      "VehicleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "transport_company_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transport Company Id"
          },
          "manufacturer": {
            "type": "string",
            "title": "Manufacturer"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "variant": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "body_type": {
            "type": "string",
            "title": "Body Type"
          },
          "gvw_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gvw Kg"
          },
          "payload_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Kg"
          },
          "tonnage_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tonnage Class"
          },
          "length_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Length Mm"
          },
          "width_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width Mm"
          },
          "height_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Mm"
          },
          "wheelbase_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wheelbase Mm"
          },
          "loading_length_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loading Length Mm"
          },
          "loading_width_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loading Width Mm"
          },
          "loading_height_mm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loading Height Mm"
          },
          "engine_cc": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Cc"
          },
          "engine_cylinders": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Cylinders"
          },
          "power_hp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Power Hp"
          },
          "torque_nm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Torque Nm"
          },
          "fuel_tank_ltr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuel Tank Ltr"
          },
          "mileage_kmpl": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mileage Kmpl"
          },
          "emission_norm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emission Norm"
          },
          "axle_config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Axle Config"
          },
          "tyres": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tyres"
          },
          "price_ex_showroom": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price Ex Showroom"
          }
        },
        "type": "object",
        "required": [
          "id",
          "transport_company_id",
          "manufacturer",
          "model_name",
          "variant",
          "category",
          "body_type",
          "gvw_kg",
          "payload_kg",
          "tonnage_class",
          "length_mm",
          "width_mm",
          "height_mm",
          "wheelbase_mm",
          "loading_length_mm",
          "loading_width_mm",
          "loading_height_mm",
          "engine_cc",
          "engine_cylinders",
          "power_hp",
          "torque_nm",
          "fuel_tank_ltr",
          "mileage_kmpl",
          "emission_norm",
          "axle_config",
          "tyres",
          "price_ex_showroom"
        ],
        "title": "VehicleResponse",
        "description": "Vehicle model response schema"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}